Add trackpad and mouse scroll functionality
This commit adds two event listeners and two new functions that handles the
scrolling functionality for trackpad and mouse wheels devices. Inside
userScrollInput() e.deltaY -+ 50 vertical scroll amount was chosen as
it seems to be a good optimised number for older and more modern trackpad
devices, and helps to prevent a double scrolling bug. isScrollUp() returns
true if the user is scrolling up, and false if the user is scrolling down.