Add touchscreen scroll
The main purpose of this commit was to add the touchscreen scroll
functionality. The touchscreen scroll works by using a touchstart and
a touchmove event that calls two separate functions. These functions
store the current value of the touch position on the X and Y axis.
These results are then later compared to determine the direction
of the user's scroll. addActiveClassToSelectors() was also
replaced with addActiveClassToList() to improve the code's
reusability. The updateState() function was also modified to fix
a undefined error when options.pageSelector was equal to false.