Add scroll speed options
The purpose of this commit was to add the scroll speed option. It works
by calling a function that sets the speed by changing the wrapper's CSS
transition duration. If the user does not set a speed option, it is
default to 1000ms. If a user sets a speed it overrides that value with
the value the user sets.