Fix IOS and MacOS Safari Issues
This commit solves two CSS Safari related issues with the scrolling.
The first fix prevents IOS safari users from scrolling down to refresh
the page. This was a problem as it stopped users from swiping
down to scroll. The second fix stopped a white flickering / flash
from appearing when the user would scroll on MacOS Safari.
Please see the comments in the CSS files for more information
on the bug fixes.